requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.30.6
pydantic==2.11.7
SQLAlchemy==2.0.32
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-ulid==2.7.0
python-dotenv==1.0.1
pydantic-settings==2.4.0
openai
requests
supabase==2.18.1
storage3==0.12.1